refactor(codegen): move the #4203 reader arm into the leaf module ✓
Keeps the ThisKeyword driver in expressions.ts to a single call, per the
LOC-regrowth ratchet: the body-swap and the marker discrimination now live
in explicit-null-receiver.ts.
