+++ src/codegen/context/types.ts
@@ -2095,12 +2095,9 @@ export interface CodegenContext {
    */
   currentThisGlobalIdx: number;
   /**
-   * (#4203) Absolute Wasm global index of the immutable `__this_explicit_null`
-   * marker — the non-null `$AnyValue` that spells "the caller passed `null` as
-   * the receiver", as distinct from `__current_this`'s `ref.null.extern` "no
-   * receiver installed". §10.4.3 gives those two states different answers in a
-   * strict callee, and one value cannot carry both. `undefined` = not
-   * reserved; see `explicit-null-receiver.ts`.
+   * (#4203) Global index of the immutable `__this_explicit_null` marker: "the
+   * caller passed `null`", as distinct from `__current_this`'s `ref.null.extern`
+   * "no receiver installed". See `explicit-null-receiver.ts`.
    */
   explicitNullThisGlobalIdx?: number;
   /** Mutable i32 hand-off used by ES5 Function `caller` poison semantics. */
